refactor voyage types
