+++ applications/vessel-history/src/types/voyage.ts
@@ -1,0 +1,19 @@
+import { ActivityEvent } from 'types/activity'
+
+export enum EventTypeVoyage {
+  Voyage = 'voyage',
+}
+
+export interface Voyage {
+  from?: ActivityEvent
+  to?: ActivityEvent
+  type: EventTypeVoyage
+  start: number
+  end: number
+  timestamp: number
+}
+
+export interface RenderedVoyage extends Voyage {
+  status: 'collapsed' | 'expanded'
+  visible: boolean
+}
